requirements.txt
asgiref==3.8.1
Django==5.2
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
mccabe==0.7.0
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
sqlparse==0.5.3
